+++ src/source/done.rs
@@ -73,6 +73,13 @@ where
     }
 }
 
+impl<I, F> ExactSizeIterator for Done<I, F>
+where
+    I: Source + ExactSizeIterator,
+    F: FnMut(&mut I),
+{
+}
+
 impl<I, F> Source for Done<I, F>
 where
     I: Source,